requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.4.0
cffi==1.16.0
channels==4.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.5
daphne==4.1.2
Django==5.0.4
django-background-task==0.1.8
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-compat==1.0.15
django-environ==0.11.2
django-redis==5.4.0
django-timezone-field==6.1.0
django4-background-tasks==1.2.9
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
kombu==5.3.7
mysqlclient==2.2.4
paho-mqtt==1.6.0
prompt-toolkit==3.0.43
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pymodbus==3.6.8
pyModbusTCP==0.2.1
pyOpenSSL==24.1.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.3
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.2
